desktop/requirements.txt
requests
speechrecognition
pyaudio
TTS
pygame
pyautogui
pyperclip
edge-tts
selenium
webdriver-manager
plyer
python-dotenv
psutil
openwakeword
numpy
onnxruntime
pycaw
comtypes
